service and controler fixed

+++ src/main/java/com/luan/baozi_store/controller/ClienteController.java
@@ -4,8 +4,11 @@ import com.luan.baozi_store.model.Cliente;
 import com.luan.baozi_store.service.ClienteService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import java.util.List;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import java.net.URI;
+import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/clientes")
@@ -18,29 +21,41 @@ public class ClienteController {
     }
 
     @GetMapping
-    public List<Cliente> findAll(){
+    public List<Cliente> findAll() {
         return clienteService.findAll();
     }
 
-
     @GetMapping("/{id}")
-    public ResponseEntity<Cliente> findById(@PathVariable Long id){
-        return clienteService.findById(id);
+    public ResponseEntity<Cliente> findById(@PathVariable Long id) {
+        Optional<Cliente> cliente = clienteService.findById(id);
+
+        if (cliente.isPresent()) {
+            return ResponseEntity.ok(cliente.get());
+        }
+
+        return ResponseEntity.notFound().build();
     }
 
     @PostMapping
-    public ResponseEntity<Cliente> saveCliente(@RequestBody Cliente cliente){
-         return clienteService.saveCliente(cliente);
+    public ResponseEntity<Cliente> saveCliente(@RequestBody Cliente cliente) {
+        Cliente clienteSalvo = clienteService.save(cliente);
+
+        URI location = ServletUriComponentsBuilder
+                .fromCurrentRequest()
+                .path("/{id}")
+                .buildAndExpand(clienteSalvo.getId())
+                .toUri();
+
+        return ResponseEntity.created(location).body(clienteSalvo);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteById(@PathVariable Long id){
-        return clienteService.deleteById(id);
+    public ResponseEntity<Void> deleteById(@PathVariable Long id) {
+        if (clienteService.existsById(id)) {
+            clienteService.deleteById(id);
+            return ResponseEntity.noContent().build();
+        }
 
+        return ResponseEntity.notFound().build();
     }
-
-
 }
-
-
-
